requirements.txt
openai==1.63.0
streamlit==1.42.0
python-dotenv==1.0.1
tqdm==4.67.1
pytest==8.3.4
